pyproject.toml
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "airavata-cerebrum"
version = "0.1.2"
description = "Apache Airavata Cerebrum : Integrated Data-driver Modeling for Computational Neuroscience"
authors = ["Sriram P Chockalingam <srirampc@gatech.edu>"]
maintainers = ["Apache Airavata Developers <dev@apache.airavata.org>"]
readme = "README.md"
homepage = "https://github.com/apache/airavata-cerebrum"
license = "Apache-2.0"
keywords = [
    "Data-driven",
    "Computational Neuroscience",
    "Neuroscience Modeling",
]

[tool.poetry.dependencies]
python = "^3.10"
abc_atlas_access = { git = "https://github.com/alleninstitute/abc_atlas_access" }
aisynphys = { git = "https://github.com/alleninstitute/aisynphys" }
awitree = { git = "https://github.com/srirampc/awitree" }
codetiming = { git = "https://github.com/srirampc/codetiming" }
anywidget = "^0.9"
allensdk = "^2.16"
altair = "^6.0"
anndata = "^0.10"
airavata-python-sdk = "^1.1.6"
bmtk = "^1.1"
duckdb = "^1.2"
fastexcel = "^0.13"
flake8 = "^7.1"
intervaltree = "^3.1.0"
ipywidgets = "^8.1"
ipytree = "^0.2"
jupyterlab = "^3.5"
leidenalg = "^0.10"
marimo = "^0.11"
matplotlib = "^3.7"
neuroanalysis = "^0.0.7"
numba = "^0.60"
numpy = "^1.23"
openpyxl = "^3.1"
parse = "^1.20"
pandas = "^1.5"
polars = "1.26"
pillow = "^10.4"
psutil = "^7.0"
pydantic = "^2.7"
python-jsonpath = "^1.2.0"
pyyaml = "^6.0"
requests = "^2.31"
scikit-learn = "^1.5"
scanpy = "^1.9"
scipy = "^1.9"
sqlalchemy = "^2.0"
sqlglot = "^26.0"
statsmodels = "^0.14"
traitlets = "^5.1"
tqdm = "^4.66"

[tool.setuptools.packages.find]
include = ["airavata_cerebrum"]

[tool.pyright]
include = ["airavata_cerebrum", "tests", "resources"]
exclude = ["**/__pycache__"]
# reportOptionalMemberAccess = false,
# reportOptionalSubscript = false,
# reportGeneralTypeIssues = false,

[tool.black]
line-length = 88
target-version = ["py310"]
